requirements.txt
numpy>=1.21.2
requests>=2.25.1
tqdm>=4.59.0
termcolor>=1.1.0
opencv-python~=4.5.5.64
pillow>=8.3.2
pycocotools>=2.0
visualdl>=2.2.0
paddlelite>=2.11
paddle2onnx>=0.9.8